Correct iteration limit when evaluating Haydock sums
In EpsL and in GreenP I evaluated the continued fractions up to but
excluding $self->nr->iteration.  This created an error in 1D where
there are only a few Haydock coefficients. In general, it should be
inoffensive, but not always. In 1D it led to serious errors. Should
continue checking.