Factor reorthogonalization. Add new Geometry instance.
Made two new roles to reorthogonalize the states for the cases of real
and of complex Haydock coefficients. Modified OneH and AllH for the
LE/NR2 and NE/S cases. Made new test suites and adapted the old ones.

Added Photonic::Geometry::FromEpsilon to get a geometry object from
the dielectric (scalar) function epsilon(\vec r). The generalization
to tensor is simple.